requirements.txt
psycopg2-binary
flask
Flask-SQLAlchemy
gunicorn
requests
pytz
python-decouple
celery
celery[redis]
redis
pathlib2
python-decouple
pydantic
google-api-python-client
google-auth
google-cloud-compute